requirements.txt
python-telegram-bot==13.11
git+https://github.com/openai/whisper.git 
boto3
